+++ src/plugins/widget/GraphPanelWidget/src/graphpanelplotwidget.h
@@ -212,7 +212,7 @@ public:
     quint64 dataSize() const;
 
     QwtSeriesData<QPointF> *data(int pRun = -1) const;
-    void setData(double *pDataX, double *pDataY, int pSize, int pRun = -1);
+    void setData(double *pDataX, double *pDataY, quint64 pSize, int pRun = -1);
 
     QRectF boundingRect();
     QRectF boundingLogRect();
@@ -489,7 +489,7 @@ public:
                  bool pEmitSignal = true, bool pForceXAxisSetting = false,
                  bool pForceYAxisSetting = false);
 
-    bool drawGraphFrom(GraphPanelPlotGraph *pGraph, int pFrom);
+    bool drawGraphFrom(GraphPanelPlotGraph *pGraph, quint64 pFrom);
 
     GraphPanelPlotWidgets neighbors() const;
